requirements.txt
gym >= 0.26
numpy==1.24.3
jax[cuda12]==0.4.24
distrax==0.1.5
flax==0.8.0
ml_collections >= 0.1.0
tqdm >= 4.60.0
chex==0.1.86
optax==0.2.2
orbax-checkpoint==0.4.4
absl-py >= 0.12.0
scipy >= 1.6.0
wandb >= 0.12.14
tensorflow==2.15.0
tensorflow-probability==0.23.0
einops >= 0.6.1
imageio >= 2.31.1
moviepy >= 1.0.3
pre-commit == 3.3.3
overrides
ml-dtypes==0.2.0
d4rl
cython < 3
patchelf
gymnasium
omegaconf==2.3.0
hydra-core==1.3.2
opencv-python==4.7.0.72
dm_control<=1.0.20
mujoco<=3.1.6
